PDFBOX-4892: LGTM fix
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1897306 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/graphics/image/CCITTFactory.java
@@ -302,7 +302,7 @@ public final class CCITTFactory
             }
 
             // Relocate to the first set of tags
-            int address = readlong(endianess, reader);
+            long address = readlong(endianess, reader);
             reader.seek(address);
     
             // If some higher page number is required, skip this page's tags,
